requirements.txt
gitpython
pytest
pyyaml
